Avoid `--chmod=755` in Dockerfile.
It seems `--chmod` is only support with a docker add-on. This causes home assistant to give this error:

```
Failed to install add-on

the --chmod option requires BuildKit. Refer to https://docs.docker.com/go/buildkit/
to learn how to build images with BuildKit enabled
```

Since we are calling `sh` as the `ENTRYPOINT`, it doesn't need to be +x anyway.